refactor: calculator 패키지 이동
