+++ src/main/kotlin/calculator/mathcalculator/Calculator.kt
@@ -1,4 +1,4 @@
-package study.caculator.mathcalculator
+package calculator.mathcalculator
 
 interface Calculator {
     fun calculate(x: Int, y: Int): Int
