+++ src/main/kotlin/calculator/mathcalculator/DivideCalculator.kt
@@ -1,4 +1,4 @@
-package study.caculator.mathcalculator
+package calculator.mathcalculator
 
 class DivideCalculator : Calculator {
     override fun calculate(x: Int, y: Int): Int {
