+++ src/main/kotlin/calculator/mathcalculator/MultiplyCalculator.kt
@@ -1,4 +1,4 @@
-package study.caculator.mathcalculator
+package calculator.mathcalculator
 
 class MultiplyCalculator : Calculator {
     override fun calculate(x: Int, y: Int): Int {
