+++ src/main/kotlin/calculator/mathcalculator/PlusCalculator.kt
@@ -1,4 +1,4 @@
-package study.caculator.mathcalculator
+package calculator.mathcalculator
 
 class PlusCalculator : Calculator {
     override fun calculate(x: Int, y: Int): Int {
