+++ src/test/kotlin/calculator/TextCalculatorTest.kt
@@ -1,4 +1,4 @@
-package study.caculator
+package calculator
 
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatCode
